requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
torch>=1.9.0
plotly>=5.1.0
streamlit>=0.84.0
pyyaml>=5.4.1
pytest>=6.2.0